OpenAI Gym simulation environment
This Jupyter file contains an implementation of a GUI simulation of a Quadcopter environment. The pid controllers are dynamically tuned via Reinforcement Learning. Currently the simulation works but the RL algorithm needs to be adjusted. Additionally we can use this simulation environment for comparing our simulink tuned pid controllers with the RL algorithm (or other tuning methods that we performed).